requirements.txt
certifi==2020.6.20
cycler==0.10.0
grpcio==1.30.0
grpcio-tools==1.30.0
kiwisolver==1.2.0
matplotlib==3.3.1
numpy==1.19.0
opencv-python==4.4.0.42
Pillow==7.2.0
protobuf==3.12.2
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
